pyproject.toml
[project]
name = "web-content-summarizer"
version = "0.1.0"
description = "Minimal web content summarizer pipeline"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "httpx>=0.28.1",
    "openai>=1.68.2",
    "python-dotenv>=1.0.1",
    "pydantic>=2.11.2",
    "trafilatura>=2.0.0",
]

[project.scripts]
web-content-summarizer = "web_content_summarizer.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true
